fastapi_observer/tracing/setup.py
from fastapi import FastAPI
from opentelemetry import trace
from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
from opentelemetry.sdk.resources import SERVICE_NAME, Resource
from opentelemetry.sdk.trace import TracerProvider


def setup_otel(app: FastAPI, service_name: str) -> None:
    trace.set_tracer_provider(
        TracerProvider(
            resource=Resource.create({SERVICE_NAME: service_name}),
        )
    )

    FastAPIInstrumentor.instrument_app(
        app, excluded_urls="health", exclude_spans=["send", "receive"]
    )
