pyproject.toml
[tool.poetry]
name = "fastapi-observer"
version = "0.0.1"
description = "Observer for FastAPI"
authors = ["Daniil Smirnov <skeal.smirnov2004@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = [
    {version = ">=0.112.4"},
]
opentelemetry-api = "^1.30.0"
opentelemetry-sdk = "^1.30.0"
opentelemetry-instrumentation-fastapi = "0.51b0"
structlog = "24.4.0"
starlette = ">=0.12.2"
prometheus_client = ">=0.12"

[tool.poetry.group.dev.dependencies]
autoflake = "2.2.0"
black = "23.7.0"
flake8 = "6.0.0"
flake8-black = "0.3.6"
flake8-bugbear = "23.7.10"
flake8-coding = "1.3.2"
flake8-comprehensions = "3.14.0"
flake8-debugger = "4.1.2"
flake8-eradicate = "1.5.0"
flake8-executable = "2.1.3"
flake8-isort = "6.0.0"
flake8-logging-format = "0.9.0"
flake8-pie = "0.16.0"
flake8-pytest-style = "1.7.2"
flake8-quotes = "3.3.2"
flake8-simplify = "0.20.0"
isort = "5.12.0"
mypy = "1.15.0"
setuptools = "78.1.0"

[tool.black]
line-length = 100
skip-string-normalization = true

[tool.isort]
line_length = 100
combine_as_imports = true
order_by_type = true
length_sort_straight = true
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3
known_local_folder = ["fastapi_observer"]
extend_skip = ["*pypoetry*"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"